StreamliAPP.py
import os
import json
import pandas as pd
import traceback
import streamlit as st
from dotenv import load_dotenv
from src.mcqgenerator.utiles import read_file,get_table_data
from src.mcqgenerator.logger import logging
from langchain.llms import OpenAI
from langchain.prompts import PromptTemplate
from langchain.chains import LLMChain
from langchain.chains import SequentialChain
from langchain.callbacks import get_openai_callback
import PyPDF2
