requirement.txt
OpenAI 
Langchain 
streamlit
python-dotenv
pypdf2

-e.
